move usage style helper to lib/styles
